requirements.txt
black==22.1.0
click==8.0.3
colorama==0.4.4
commonmark==0.9.1
mypy-extensions==0.4.3
numpy==1.22.2
pandas==1.4.0
pathspec==0.9.0
platformdirs==2.4.1
psycopg2-binary==2.9.3
Pygments==2.11.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
rich==11.2.0
six==1.16.0
snakeviz==2.1.1
tabulate==0.8.9
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
nltk~=3.7
tensorflow~=2.8.0
keras~=2.8.0
scikit-learn~=1.0.2